resolution SFINAE

+++ test.cpp
@@ -25,7 +25,7 @@ int main() {
 	std::cout << (paire1 > paire2) << std::endl; 
 	std::cout << (paire1 == paire2) << std::endl; 
 
-	ft::Vector<int> vec(static_cast<size_t>(4), static_cast<int>(3));
+	ft::Vector<int> vec(4,3);
 	std::cout << paire1.first << std::endl;
 	ft::Vector<int> vec2(vec);
 	vec2.resize(2, 2);
